# backend/docs/swagger.yaml
basePath: /
definitions:
...
  httperrors.ErrorResponse:
    properties:
      details:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
  repository.Chat:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      metadata:
        items:
          type: integer
        type: array
      title:
        type: string
      type:
        $ref: '#/definitions/repository.ChatType'
      updated_at:
        type: string
    type: object
  repository.ChatType:
    enum:
    - private
    - group
    - channel
    type: string
    x-enum-varnames:
    - ChatTypePrivate
    - ChatTypeGroup
    - ChatTypeChannel
  repository.Invitation:
    properties:
      chat_id:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  repository.Message:
    properties:
      chat_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      reply_message_id:
        type: string
      sender_id:
        type: string
      topic_id:
        type: string
      updated_at:
        type: string
    type: object
  repository.Topic:
    properties:
      avatar_url:
        type: string
      chat_id:
        type: string
      creted_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      title:
        type: string
      type:
        $ref: '#/definitions/repository.TopicType'
      updated_at:
        type: string
    type: object
  repository.TopicType:
    enum:
    - text_topic
    - voice_topic
    type: string
    x-enum-varnames:
    - TopicTypeTextTopic
    - TopicTypeVoiceTopic
  repository.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      last_seen_at:
        type: string
      metadata:
        items:
          type: integer
        type: array
      password_hash:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  services.CreateMessageDTO:
    properties:
      content:
        example: Hello, world!
        type: string
      reply_message_id:
        type: string
    type: object
...
  services.InvitationInfo:
    properties:
      chat:
        $ref: '#/definitions/repository.Chat'
      creator:
        $ref: '#/definitions/repository.User'
      invitation:
        $ref: '#/definitions/repository.Invitation'
    type: object
...
  services.WebPushNotificationDTO:
    properties:
      body:
        type: string
      title:
        type: string
    type: object
  services.WebPushSubscriptionDTO:
    properties:
      endpoint:
        type: string
      keys:
        properties:
          auth:
            type: string
          p256dh:
            type: string
        type: object
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for a messenger application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Messenger API
  version: "1.0"
paths:
...
  /invitation/{id}:
    delete:
      consumes:
      - application/json
      description: Reject or delete a chat invitation.
      parameters:
      - description: Invitation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reject invitation
      tags:
      - invitations
  /invitation/{id}/accept:
    post:
      consumes:
      - application/json
      description: Accept a chat invitation and join the chat.
      parameters:
      - description: Invitation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Accept invitation
      tags:
      - invitations
  /invitation/{id}/info:
    get:
      consumes:
      - application/json
      description: Returns information about a chat invitation by its ID.
      parameters:
      - description: Invitation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.InvitationInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get invitation info
      tags:
      - invitations
  /pubsub/push/notify:
    post:
      consumes:
      - application/json
      description: Send a push notification to all subscribed users.
      parameters:
      - description: Notification details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.WebPushNotificationDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
      summary: Send push notification
      tags:
      - pubsub
  /pubsub/push/pubkey:
    get:
      description: Returns the VAPID public key needed for Web Push subscription.
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
      summary: Get VAPID public key
      tags:
      - pubsub
  /pubsub/push/subscribe:
    post:
      consumes:
      - application/json
      description: Save a Web Push subscription for the authenticated user.
      parameters:
      - description: Push subscription details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.WebPushSubscriptionDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Subscribe to push notifications
      tags:
      - pubsub
  /pubsub/sse:
    get:
      description: Subscribe to Server-Sent Events for real-time chat updates.
      produces:
      - text/event-stream
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: SSE stream
      tags:
      - pubsub
  /topics/{id}:
    get:
      consumes:
      - application/json
      description: Returns a topic by ID.
      parameters:
      - description: Topic ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repository.Topic'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a topic by ID
      tags:
      - topics
  /topics/{id}/messages:
    get:
      consumes:
      - application/json
      description: Returns a list of all messages in a topic.
      parameters:
      - description: Topic ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repository.Message'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all messages in a topic
      tags:
      - messages
    post:
      consumes:
      - application/json
      description: Creates a new message in a topic.
      parameters:
      - description: Topic ID
        in: path
        name: id
        required: true
        type: string
      - description: Message details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.CreateMessageDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repository.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new message in a topic
      tags:
      - messages
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by your JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"